Remove integration tests
We now expect that integration tests lie closer to the consuming
libraries. They're still _really_ important, however a lot of these
tests rely on the environment in which they're run. In particular,
Masonry's tests are highly dependent on the actual things you put
inside Masonry.

The tests themselves still exist internally within Pinterest so no
coverage is actually being lost. The upside is that we can
significantly simply the project setup here.