Removes integration tests (#439)
* Remove integration tests

We now expect that integration tests lie closer to the consuming
libraries. They're still _really_ important, however a lot of these
tests rely on the environment in which they're run. In particular,
Masonry's tests are highly dependent on the actual things you put
inside Masonry.

The tests themselves still exist internally within Pinterest so no
coverage is actually being lost. The upside is that we can
significantly simply the project setup here.

* update yarn.lock

* upgrade rollup and rm more deps

* test against dubnium

* update react-scripts

* remove rollup progress to unbreak build

* downgrade a couple of deps to exact versions for CRA