Avatar: Add accessibilityLabel (#961)
We currently always use the name for the aria-label (or svg title), which is a decent default but not always desired. For example, when using an Avatar next to the user's name, a screen reader will read the name twice (once for the Avatar's aria-label, once for the text content). This also allows for more flexibility if the aria-label should be something instead of the user's name.