[Box] Fix types (#986)
We found with #966 that Box's props are effectively untyped - any. This lead to an incident where somebody was incorrectly setting zIndex on Box, but Flow didn't catch it. This fixes the Flow suppressions on Box.

## Test Plan

I know this changed worked because Flow immediately found type errors in the DatePicker due to role and zIndex being (incorrectly) marked as required props. This diff also fixes those.