Merge: Better titles and obey branch protections (#1035)
I noticed that PRs (like #1032) weren't getting merged if they weren't up to date with master. This is because they have a special status check that we have set to fail if the branch isn't up to date with master (even if the rebase is clean). Changing the "base=master" value fixes this and has Mergify rebase PRs automatically and also has it obey that branches status checks (that's why I was able to remove the explicit status-check check).

Note: I'll disable the requirement of PRs being up to date with master once this lands.

I also noticed that the commit description was pretty useless because people are used to updating the PR description and not have meaningful commit names because we squash them anyway. 809000b is an example. I've also changed the commit_message var to reformat the message from the title + body of the PR which is more in line with what folks are used to.
Test Plan

I've verified this config on a bunch of open PRs and it's working pretty well.