AbstractEventHandler type (#1045)
We've discussed this a little bit elsewhere - having to repeat the standard event format is unusual and error prone. This starts to consolidate all the event handlers to have the same enforced type. This also opens up an easy refactoring to change all event handlers to return `mixed` instead of `void` to make them place nicer with async functions.

I've updated most events across Gestalt to use it and also added a few type refinements where they were missing (filling in `SyntheticEvent<>` holes) The documentation doesn't need to change because the underlying structure is still the same.

- [x] Document `AbstractEventHandler` in a general section talking about event handlers and _why_ we have the `{event}` object instead of the normal callback format.

## Test Plan

Flow