Docs: Fix Header to remove sticky header on reduced window size (#1074)
The `<Sticky />` header causes issues in our docs. It was introduced as part of #952 /cc @AlbertCarreras 

However,  it is convenient to have a sticky header on Gestalt Docs due to the amount of scrolling that is required when reading the Docs. 

A proposed solution is to keep the sticky header on full windows and remove it when windows downsize to a size that doesn't allow the correct display of Modals.

## Before
![Screen Shot 2020-07-23 at 2 22 06 PM](https://user-images.githubusercontent.com/127199/88340179-34e71c00-ccf0-11ea-9f96-2d6fdb36c286.png)

## After
![Kapture 2020-07-23 at 18 19 17](https://user-images.githubusercontent.com/10593890/88353092-12b1c600-cd11-11ea-8344-f3712060a9fa.gif)

<!--
What is the purpose of this PR?

* What is the context surrounding this PR? Include links if possible.
* What kind of feedback do you want?
* Have you [formatted the PR title](https://github.com/pinterest/gestalt/#releasing)? `ComponentName: Description`
-->

## Test Plan

<!--
How can reviewers verify this is good to merge?

* Is it tested?
* Is it accessible?
* Is it documented?
* Have you involved other stakeholders (such as a Pinterest Designer)?
-->