Docs: enable flow types (#1173)
Enable flow types on our Gestalt docs. Improves type safety and caught a couple of actual issues.

We previously did not enable these because of https://github.com/facebook/flow/issues/4015. It does mean we might have to restart the docs server when we make a change to Gestalt.

## Test Plan

`yarn flow` passes + checks the docs.