Internal: fix `flowtype/require-exact-type` errors (#1228)
We added the linter `flowtype/require-exact-type` in #946, but we disable the linter in several places. 

For the same reasons we added the `flowtype/require-exact-type` linter, disabling these errors is actually very bad because the downstream projects do not really respect the exact-by-default logic and would still treat `{ }` as inexact. We should avoid using `{ }` to flow running in downstream projects correctly know whether object type is exact and inexact without any ambiguity. 

This PR fixes all linter errors. Note that in some cases we actually meant to use inexact types. I have made sure that they are correctly typed as inexact.

## Test Plan

Flow