Table: Add onExpand function prop to RowExpandable (#1263)
## Summary
The purpose of this PR is to add the onExpand function prop to the Table.RowExpandable component. 

An example use case for this prop would be if a user wanted to request an API call **only** when the row is expanded. 

## Test Plan
I used the docs and created more tests to make sure the function prop works as needed.