Internal: Enable the eslint rule flowtype/no-mutable-array (#1321)
Enable the eslint rule [`flowtype/no-mutable-array`](https://github.com/gajus/eslint-plugin-flowtype#eslint-plugin-flowtype-rules-no-mutable-array)

> General reasons for using immutable data structures:
> - They are simpler to construct, test, and use
> - They help to avoid temporal coupling
> - Their usage is side-effect free (no defensive copies)
> - Identity mutability problem is avoided
> - They always have failure atomicity
> - They are much easier to cache

When a prop/argument of a component/function is typed as the mutable `Array`, one cannot pass a variable that is typed as a read-only array `$ReadOnlyArray`. This PR will make it possible for projects that use Gestalt to pass `$ReadOnlyArray` data structure to Gestalt components.