Layer: fix a bug where Layer ummounts children on rerender when zIndex changes (#1327)
This fixes #1326.

When we pass `zIndex` indexable object and the reference is changed, the cleanup routine of `useEffect` hook, which depends on `zIndex` would be fired and would replace the entire children tree with a new one. I believe this is an oversight when we added `zIndex` support in #1223. I am fixing this bug by:

1. make sure the `useEffect` hook which appends/removes a div element only runs on mount/unmount.
2. the zIndex is handled by a separate `useEffect` after the `useEffect` hook above so that it (1) initiailizes the zIndex after the div has been created on mount, and (2) updates zIndex-related attributes of the div element whenever the value of the zIndex prop changes.

I also removed the documentation where we encourage people to avoid rerendering. Avoid rerendering should only be a perf optimization and should not be used as a semantic guarentee.