Typeahead: add zIndex prop to support component in Modals with zIndex (#1350)
Typeahead didn't include zIndex prop to support component within Modals with existing zIndex. Other components as Dropdown do include zIndex to support these cases.

This PR implements zIndex prop in Typeahead and adds an example to the Docs

![image](https://user-images.githubusercontent.com/10593890/105769302-a2883500-5f2b-11eb-87df-f3e460675609.png)

Without the zIndex, the Typeahead is overlayed by the Modal zIndex.

![image](https://user-images.githubusercontent.com/10593890/105769590-06aaf900-5f2c-11eb-8ff1-b403b357febc.png)

NOTE: This fix allows Typeahead in Modals with zIndex but scrollable Modals will present Flyout positioning issue.  To be fixed with the upcoming ScrollableBox component.
<!--
What is the purpose of this PR?

* What is the context surrounding this PR? Include links if possible.
* What kind of feedback do you want?
* Have you [formatted the PR title](https://github.com/pinterest/gestalt/#releasing)? `ComponentName: Description`
-->

## Test Plan

<!--
How can reviewers verify this is good to merge?

* Is it tested?
* Is it accessible?
* Is it documented?
* Have you involved other stakeholders (such as a Pinterest Designer)?
-->