Box: Remove marginLeft and marginRight props (#1363)
These props have been recommended against in Pinboard since we introduced the RTL-friendly marginStart and marginEnd alternatives. Rather than rely on a lint rule, this PR removes the props altogether. This seems like a better DX, as IDEs will no longer suggest those props, and the dev can always use dangerouslySetInlineStyle if they really, really need a specific direction.

## Test Plan

This PR replaces 99% of marginLeft/marginRight strings in the repo; the remaining ones are in the lint rule against them, and the docs for that lint rule. All existing usages (both in src code and docs) should be properly updated.