Updated the design section of the page "How to work with us"
Edited the design section of the page "How to work with us," making the content simpler and easier on the eyes. In addition, I deleted a few lines and added new ones.
Reason: we explain the whole process of contribution during our office hours; there isn't a need to make it too granular on the docs, as it is causing a long scrolling and heavy content. Trying to make it more digestible.