Docs: Update legacy banner content; hide button [UXE-202]
What:

In the legacy site’s banner, hide the “Explore Gestalt 2.0” button and
change the text to: “Information might be outdated. Stay tuned for
Gestalt’s new documentation website coming soon.”

Why:

The new button links back to the current page, which is confusing and
messes up browser history. We’ll re-enable this button when the new
website content is ready.